Revert "Template analytics"
