+++ .env
@@ -1,3 +1,3 @@
-NEXT_PUBLIC_BASE_URL =https://stage.apixyzfolio.com
+NEXT_PUBLIC_BASE_URL =https://prod.apixyzfolio.com
 NEXT_PUBLIC_GOOGLE_CLIENT_ID =19639908448-pf82heavbk4rsa7bklm26tg8179nlc0u.apps.googleusercontent.com
 NEXT_PUBLIC_BASE_DOMAIN=localhost
