+++ package.json
@@ -28,10 +28,8 @@
     "@feedefy/react": "^1.0.0",
     "@react-oauth/google": "^0.12.1",
     "@tanstack/react-query": "^5.56.2",
-    "array-move": "^4.0.0",
     "axios": "^1.7.7",
     "browser-image-compression": "^2.0.2",
-    "chart.js": "^4.4.6",
     "compressorjs": "^1.2.1",
     "dayjs": "^1.11.13",
     "dompurify": "^3.1.7",
@@ -44,10 +42,8 @@
     "next": "14.2.11",
     "next-themes": "^0.3.0",
     "react": "^18",
-    "react-chartjs-2": "^5.2.0",
     "react-circular-progressbar": "^2.1.0",
     "react-dom": "^18",
-    "react-easy-sort": "^1.6.0",
     "react-medium-image-zoom": "^5.2.10",
     "react-select": "^5.8.1",
     "react-sortable-hoc": "^2.0.0",
@@ -63,7 +59,6 @@
     "@tanstack/react-query-devtools": "^5.56.2",
     "eslint": "^8",
     "eslint-config-next": "14.2.11",
-    "file-loader": "^6.2.0",
     "postcss": "^8",
     "sass": "^1.79.3",
     "tailwindcss": "^3.4.1"
