+++ src/components/ProjectCard.jsx
@@ -53,7 +53,7 @@ export default function ProjectCard({
             whileHover="hover"
             variants={imageVariants}
             loading="lazy"
-            fetchpriority="high"
+            fetchPriority="high"
             decoding="async"
             // onLoad={() => setImageLoaded(true)}
           />
@@ -123,7 +123,7 @@ export default function ProjectCard({
 const DragHandle = SortableHandle(() => (
   <div
     onClick={(e) => e.stopPropagation()}
-    className="!px-[24.5px] !cursor-grab py-[19px] transition-shadow duration-500 ease-out bg-project-card-reorder-btn-bg-color border-project-card-reorder-btn-bg-color hover:border-project-card-reorder-btn-bg-hover-color hover:bg-project-card-reorder-btn-bg-hover-color rounded-2xl"
+    className="!px-[24.5px] cursor-grab py-[19px] transition-shadow duration-500 ease-out bg-project-card-reorder-btn-bg-color border-project-card-reorder-btn-bg-color hover:border-primary-btn-bg-hover-color hover:bg-primary-btn-bg-hover-color rounded-2xl"
   >
     <DragIcon className="text-project-card-reorder-btn-icon-color !cursor-grab" />
   </div>
