+++ src/components/header.jsx
@@ -16,8 +16,6 @@ const Header = ({ dfToken, hideHeader }) => {
     cursor,
     setCursor,
     changeCursor,
-    changeTemplate,
-    template,
   } = useGlobalContext();
 
   // Determine header component based on dfToken and router pathname
@@ -42,8 +40,6 @@ const Header = ({ dfToken, hideHeader }) => {
         cursor={cursor}
         setCursor={setCursor}
         changeCursor={changeCursor}
-        changeTemplate={changeTemplate}
-        template={template}
       />
     );
   }
