+++ src/lib/cursor.js
@@ -1,8 +1,7 @@
 export const getCursor = (
   defaultCursor = "",
   pointer = "",
-  notAllowed = "",
-  grab = ""
+  notAllowed = ""
 ) => {
   document.documentElement.style.setProperty(
     "--cursor-default",
@@ -16,27 +15,14 @@ export const getCursor = (
     "--cursor-not-allowed",
     `url(${notAllowed}), not-allowed`
   );
-  document.documentElement.style.setProperty(
-    "--cursor-grab",
-    `url(${grab}), grab`
-  );
-  document.documentElement.style.setProperty(
-    "--cursor-grabbing",
-    `url(${grab}), grabbing`
-  );
 };
 
-export const removeCursor = () => {
+const removeCursor = () => {
   document.documentElement.style.setProperty("--cursor-default", "default");
   document.documentElement.style.setProperty("--cursor-pointer", "pointer");
   document.documentElement.style.removeProperty(
     "--cursor-not-allowed",
     "not-allowed"
-  );
-  document.documentElement.style.removeProperty("--cursor-grab", "grab");
-  document.documentElement.style.removeProperty(
-    "--cursor-grabbing",
-    "grabbing"
   );
 };
 
@@ -49,48 +35,42 @@ export const setCursorvalue = (cursorType) => {
       getCursor(
         "/assets/svgs/default1.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
     case 2:
       getCursor(
         "/assets/svgs/default2.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
     case 3:
       getCursor(
         "/assets/svgs/default3.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
     case 4:
       getCursor(
         "/assets/svgs/default4.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
     case 5:
       getCursor(
         "/assets/svgs/default5.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
     case 6:
       getCursor(
         "/assets/svgs/default6.svg",
         "/assets/svgs/pointer1.svg",
-        "/assets/svgs/stop.svg",
-        "/assets/svgs/move.svg"
+        "/assets/svgs/stop.svg"
       );
       break;
