--- src/pages/analytics.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Analytics from "@/components/analytics";
-
-function AnalyticsPage() {
-  return (
-    <div
-      className={`max-w-[890px] mx-auto py-[94px] md:py-[135px] px-2 md:px-4 lg:px-0`}
-    >
-      <Analytics />
-    </div>
-  );
-}
-
-export default AnalyticsPage;
