+++ src/styles/animation.css
@@ -20,26 +20,3 @@
 .bounce:nth-child(3) {
   animation-delay: 0.3s;
 }
-
-@keyframes bubble {
-  0%,
-  80%,
-  100% {
-    opacity: 1;
-  }
-  40% {
-    opacity: 0.5;
-  }
-}
-
-.bubble {
-  animation: bubble 1s infinite;
-}
-
-.bubble:nth-child(2) {
-  animation-delay: 0.15s;
-}
-
-.bubble:nth-child(3) {
-  animation-delay: 0.3s;
-}
