+++ src/styles/globals.scss
@@ -9,8 +9,6 @@
   --cursor-default: default;
   --cursor-pointer: pointer;
   --cursor-not-allowed: not-allowed;
-  --cursor-grab: "grab";
-  --cursor-grabbing: "grabbing";
 }
 
 *,
@@ -32,17 +30,6 @@ body {
 
 *:focus {
   outline: none;
-}
-
-/* Hide scrollbar for Chrome, Safari and Opera */
-.hide-scrollbar::-webkit-scrollbar {
-  display: none;
-}
-
-/* Hide scrollbar for IE, Edge and Firefox */
-.hide-scrollbar {
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
 }
 
 .text-input {
